Add validation

+++ src/main/java/com/example/usermanager/repository/CustomerRepository.java
@@ -4,8 +4,10 @@ import com.example.usermanager.domain.Customer;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface CustomerRepository extends MongoRepository<Customer, Long> {
+public interface CustomerRepository extends MongoRepository<Customer, BigInteger> {
 
     Customer findByLogin(String login);
 }
